Fix AMP dtype crash in PhysNet bond force (index_add_ Half/Float mismatch)
Under torch.autocast, the bond-force MLP's output (magnitude) lands in
fp16/bf16 (Linear layers are autocast-listed), while unit_dir -- built
from plain arithmetic on q -- is not autocast-listed and stays fp32.
Their product (contribution) therefore ends up fp32, but the
accumulator was allocated as hv.dtype (fp16), and index_add_ (unlike
slice assignment) requires an exact dtype match and throws:

    RuntimeError: index_add_(): self (Half) and source (Float)
    must have the same scalar type

Fix: allocate the accumulator's dtype from contribution itself, not
from hv. Verified the crash reproduces under real torch.autocast (cpu,
bfloat16 -- same op-casting rules as cuda/fp16) without the fix, and
that the full model (encoder + CAL head + all 4 losses + backward)
now runs clean under autocast across multiple batches.

+++ src/models/physnet.py
@@ -120,7 +120,13 @@ class PhysNet(nn.Module):
         contribution = magnitude * unit_dir                            # [E, 3]
 
         A = hv.shape[0]
-        f_bond = torch.zeros((A, 3), device=hv.device, dtype=hv.dtype)
+        # dtype is taken from `contribution`, not `hv`: under AMP autocast, magnitude
+        # (produced by a Linear layer) lands in fp16 while unit_dir (from plain
+        # tensor arithmetic on q, which autocast leaves alone) stays fp32, so
+        # `contribution`'s actual dtype can differ from hv.dtype. index_add_ (unlike
+        # plain slice assignment) requires an exact dtype match between self and
+        # source, so the accumulator must match whatever contribution really is.
+        f_bond = torch.zeros((A, 3), device=hv.device, dtype=contribution.dtype)
         f_bond.index_add_(0, begin_idx, contribution)
         return f_bond
